ai-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
python-dotenv==1.0.0
joblib==1.3.1
